# Shizou.WebApi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Shizou.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Images/MissingAnimePosters": {
      "put": {
        "tags": [
          "Images"
        ],
        "operationId": "ShizouHttpClient_ImagesGetMissingAnimePosters",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/Images/AnimePosters/{animeId}": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "ShizouHttpClient_ImagesGetAnimePoster",
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": { }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/Images/CreatorImages/{creatorId}": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "ShizouHttpClient_ImagesGetCreatorImage",
        "parameters": [
          {
            "name": "creatorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": { }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/Images/EpisodeThumbnails/{episodeId}": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "ShizouHttpClient_ImagesGetEpisodeThumbnail",
        "parameters": [
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": { }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}